Move common/users-client to common/users
Also, rename users.New -> users.NewClient
